rename dir and refine struct
