+++ python/nndeploy/test/op/test_activation.py
@@ -4,28 +4,23 @@ import torch
 import nndeploy
 from nndeploy.op import functional as F
 
-from nndeploy.test_utils import (
-    createTensorFromNumpy,
-    createNumpyFromTensor,
-    device_name_to_code,
-)
+from nndeploy.test.test_util import createTensorFromNumpy, createNumpyFromTensor
 
 
-class TestGlobalAveragePoolOp(unittest.TestCase):
+class TestActivation(unittest.TestCase):
 
-    def test_global_averagepool(self):
-        input_shape = [32, 64, 32, 32]
-
+    def test_relu_0(self):
+        input_shape = [32, 4, 16, 16]
+        
         np_input = np.random.random(input_shape).astype(np.float32)
-
-        torch_result = torch.nn.functional.adaptive_avg_pool2d(
-            torch.tensor(np_input),
-            output_size=(1, 1),
+        
+        torch_result = torch.nn.functional.relu(
+            torch.tensor(np_input)
         )
 
         input = createTensorFromNumpy(np_input)
 
-        nndeploy_result = F.global_averagepool(input)
+        nndeploy_result = F.relu(input)
 
         self.assertTrue(
             np.allclose(
