+++ python/nndeploy/test/op/test_global_averagepool.py
@@ -4,27 +4,24 @@ import torch
 import nndeploy
 from nndeploy.op import functional as F
 
-from nndeploy.test_utils import (
-    createTensorFromNumpy,
-    createNumpyFromTensor,
-    device_name_to_code,
-)
+from nndeploy.test.test_util import createTensorFromNumpy, createNumpyFromTensor
 
 
-class TestActivation(unittest.TestCase):
+class TestGlobalAveragePoolOp(unittest.TestCase):
 
-    def test_relu_0(self):
-        input_shape = [32, 4, 16, 16]
-        
+    def test_global_averagepool(self):
+        input_shape = [32, 64, 32, 32]
+
         np_input = np.random.random(input_shape).astype(np.float32)
-        
-        torch_result = torch.nn.functional.relu(
-            torch.tensor(np_input)
+
+        torch_result = torch.nn.functional.adaptive_avg_pool2d(
+            torch.tensor(np_input),
+            output_size=(1, 1),
         )
 
         input = createTensorFromNumpy(np_input)
 
-        nndeploy_result = F.relu(input)
+        nndeploy_result = F.global_averagepool(input)
 
         self.assertTrue(
             np.allclose(
