+++ python/nndeploy/test/test_util/__init__.py
@@ -1,0 +1,2 @@
+from .util import createTensorFromNumpy, createNumpyFromTensor
+from .util import str_to_np_data_types, device_name_to_code
