--- python/nndeploy/test_utils/__init__.py
@@ -1,2 +1,0 @@
-from .test_util import createTensorFromNumpy, createNumpyFromTensor
-from .test_util import str_to_np_data_types, device_name_to_code
